add easier examples
